add checks on re_category deletion

+++ Modules/resources/Controller/RecategoriesController.php
@@ -7,6 +7,7 @@ require_once 'Modules/core/Controller/CoresecureController.php';
 require_once 'Modules/core/Model/CoreStatus.php';
 require_once 'Modules/resources/Model/ResourcesTranslator.php';
 require_once 'Modules/resources/Model/ReCategory.php';
+require_once 'Modules/resources/Model/ReCategory.php';
 
 /**
  * 
@@ -96,8 +97,20 @@ class RecategoriesController extends CoresecureController {
     
     public function deleteAction($id_space, $id){
         $this->checkAuthorizationMenuSpace("resources", $id_space, $_SESSION["id_user"]);
-        
-        $this->categoryModel->delete($id_space, $id);
+
+        // check if category is linked to resources. If yes, deletion is not authorized => warn the user
+        $resourceModel = new ResourceInfo();
+        $linkedResources = $resourceModel->resourcesForCategory($id_space, $id);
+
+        if ($linkedResources == null || empty($linkedResources)) {
+            Configuration::getLogger()->debug("[TEST][CAT]", ["not linked to resources, deletion is authorized"]);
+            // not linked to resources, deletion is authorized
+            $this->categoryModel->delete($id_space, $id);
+        } else {
+            Configuration::getLogger()->debug("[TEST][CAT]", ["linked to resources, deletion is NOT authorized"]);
+            // not linked to resources, notify the user
+            $_SESSION["message"] = "Error: As long as this category is linked to existing resources, you can't delete it";
+        }
         $this->redirect("recategories/".$id_space);
     }
 }
